wip: improviment on tab component and it add the icon option on tab
